feat: add mdx component & single post styling
