+++ next.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  trailingSlash: true,
+}
