+++ package.json
@@ -23,9 +23,11 @@
     "autoprefixer": "^10.1.0",
     "classnames": "^2.2.6",
     "gray-matter": "^4.0.2",
+    "mdx-utils": "^0.2.0",
     "next": "10.0.3",
     "next-mdx-remote": "^1.0.1",
     "postcss": "^8.2.0",
+    "prism-react-renderer": "^1.1.1",
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "tailwindcss": "^2.0.1"
@@ -36,5 +38,9 @@
   },
   "bugs": {
     "url": "https://github.com/omarhoumz/omarhoumz.github.io/issues"
+  },
+  "devDependencies": {
+    "stylelint": "^13.8.0",
+    "stylelint-config-standard": "^20.0.0"
   }
 }
