--- src.old/components/_ui/mdx/pre-code-highlighted/pre-code-highlighted.module.css
@@ -1,37 +1,0 @@
-.preWrapper {
-  --inline-space: 1rem;
-  display: flex;
-  overflow: auto;
-  margin: 1rem calc(-1 * var(--inline-space));
-  border: 1px solid var(--blue-50);
-  border-radius: 4px;
-  overflow: hidden;
-}
-
-@media (min-width: 860px) {
-  .preWrapper {
-    --inline-space: 2rem;
-  }
-}
-
-.preStyles {
-  display: inline-block;
-  margin: 0;
-  padding: 0.5rem;
-
-  text-align: left;
-  width: 100%;
-  overflow: auto;
-}
-
-.line {
-  display: table-row;
-}
-
-.lineNo {
-  display: table-cell;
-  text-align: right;
-  padding-right: 1rem;
-  user-select: none;
-  opacity: 0.5;
-}
