--- src/components/mdx/code/code.module.css
@@ -1,7 +1,0 @@
-.code {
-  background-color: #f2f2f2;
-  border-radius: 2px;
-  padding: 1px 3px;
-  border: 1px solid #e2e6ea;
-  color: #24292e;
-}
