+++ src/components/mdx/components/components.jsx
@@ -1,0 +1,32 @@
+import React from 'react'
+import { preToCodeBlock } from 'mdx-utils'
+
+import PreCodeHighlighted from 'src/components/mdx/pre-code-highlighted/pre-code-highlighted'
+import Link from 'src/components/link/link'
+
+import styles from './components.module.css'
+
+const Pre = (preProps) => {
+  const props = preToCodeBlock(preProps)
+  // if there's a codeString and some props, we passed the test
+  if (props) {
+    return <PreCodeHighlighted {...props} />
+  }
+  return <pre {...props} />
+}
+
+const Blockquote = (props) => (
+  <blockquote className={styles.blockquote} {...props} />
+)
+
+const Code = (props) => {
+  console.log(props)
+  return <code className={styles.code} {...props} />
+}
+
+export const components = {
+  pre: Pre,
+  code: Code,
+  blockquote: Blockquote,
+  a: Link,
+}
