+++ stylelint.config.js
@@ -1,0 +1,19 @@
+module.exports = {
+  extends: ['stylelint-config-standard'],
+  rules: {
+    'at-rule-no-unknown': [
+      true,
+      {
+        ignoreAtRules: [
+          'tailwind',
+          // 'apply',
+          'variants',
+          'responsive',
+          'screen',
+        ],
+      },
+    ],
+    // 'declaration-block-trailing-semicolon': null,
+    // 'no-descending-specificity': null,
+  },
+}
